Reviewed-on: https://git.onlyoffice.com/ONLYOFFICE/onlyoffice-owncloud/pulls/49

+++ js/settings.js
@@ -295,26 +295,36 @@
     });
 
     $("#onlyofficeClearVersionHistory").click(function () {
-      $(".section-onlyoffice").addClass("icon-loading");
-
-      $.ajax({
-        method: "DELETE",
-        url: OC.generateUrl(
-          "apps/" + OCA.Onlyoffice.AppName + "/ajax/settings/history"
-        ),
-        success: function onSuccess(response) {
-          $(".section-onlyoffice").removeClass("icon-loading");
-          if (response) {
-            const message = t(
-              OCA.Onlyoffice.AppName,
-              "All history successfully deleted"
-            );
-            OC.Notification.show(message, {
-              timeout: 3,
-            });
+      OC.dialogs.confirm(
+        t(OCA.Onlyoffice.AppName, 'Are you sure you want to clear metadata?'),
+        t(OCA.Onlyoffice.AppName, 'Confirm metadata removal'),
+        (clicked) => {
+          if (!clicked) {
+            return
           }
+
+          $(".section-onlyoffice").addClass("icon-loading");
+
+          $.ajax({
+            method: "DELETE",
+            url: OC.generateUrl(
+              "apps/" + OCA.Onlyoffice.AppName + "/ajax/settings/history"
+            ),
+            success: function onSuccess(response) {
+              $(".section-onlyoffice").removeClass("icon-loading");
+              if (response) {
+                const message = t(
+                  OCA.Onlyoffice.AppName,
+                  "All history successfully deleted"
+                );
+                OC.Notification.show(message, {
+                  timeout: 3,
+                });
+              }
+            },
+          });
         },
-      });
+      )
     });
 
     $("#onlyofficeAddTemplate").change(function () {
